requirements.txt
requests
beautifulsoup4
pyzbar
Pillow